requirements.txt
Django>=4.2
djangorestframework
celery
amqp
gunicorn
django-cors-headers
